fix description of messageDisplay functions
It is possible that you don't have the message pane, but only the message list, so there is no message displayed.
I could not find out how it is possible to display multiple messages in a tab, but I guess this references the view you get in the message pane when you select more than 1 message. I don't think it is possible to display multiple messages without selecting them in the same tab, but the other way round is possible.